[QA] Enforce const-correctness

+++ src/vtpty/UnixUtils.cpp
@@ -21,7 +21,7 @@ UnixPipe::UnixPipe(unsigned flags): pfd { -1, -1 }
     if (pipe(pfd.data()) < 0)
         throw std::runtime_error { "Failed to create PTY pipe. "s + strerror(errno) };
     for (auto const fd: pfd)
-        if (!util::setFileFlags(fd, flags))
+        if (!util::setFileFlags(fd, static_cast<int>(flags)))
             break;
 #endif
 }
